DLR wait for final_stats in httpapi
In a shipment via SMPP the DLR manager waits for a PDU with one of the final statuses to remove the DLR from redis. In an HTTP submission this did not happen. Thus, DELIVRD were lost if a PDU with an ACCEPTD was sent first even though the HTTP send was made with DLR_LEVEL = 3

+++ jasmin/managers/dlr.py
@@ -313,7 +313,8 @@ class DLRLookup(object):
                 raise DLRMapNotFound('Got a DLR for an unknown message id: %s (coded:%s)' % (pdu_dlr_id, msgid))
             if len(dlr) > 0 and dlr['sc'] != connector_type:
                 raise DLRMapError('Found a dlr for msgid:%s with diffrent sc: %s' % (submit_sm_queue_id, dlr['sc']))
-
+            success_states = ['ACCEPTD', 'DELIVRD']
+            final_states = ['DELIVRD', 'EXPIRED', 'DELETED', 'UNDELIV', 'REJECTD']
             if connector_type == 'httpapi':
                 self.log.debug('There is a HTTP DLR request for msgid[%s] ...', msgid)
                 dlr_url = dlr['url']
@@ -341,8 +342,9 @@ class DLRLookup(object):
                                                                                text=pdu_dlr_text,
                                                                                method=dlr_method))
 
-                    self.log.debug('Removing HTTP dlr map for msgid[%s]', submit_sm_queue_id)
-                    yield self.redisClient.delete('dlr:%s' % submit_sm_queue_id)
+                    if pdu_dlr_status in final_states:
+                        self.log.debug('Removing HTTP dlr map for msgid[%s]', submit_sm_queue_id)
+                        yield self.redisClient.delete('dlr:%s' % submit_sm_queue_id)
             elif connector_type == 'smppsapi':
                 self.log.debug('There is a SMPPs mapping for msgid[%s] ...', msgid)
                 system_id = dlr['system_id']
@@ -354,9 +356,6 @@ class DLRLookup(object):
                 destination_addr = str(dlr['destination_addr'])
                 sub_date = dlr['sub_date']
                 registered_delivery_receipt = dlr['rd_receipt']
-
-                success_states = ['ACCEPTD', 'DELIVRD']
-                final_states = ['DELIVRD', 'EXPIRED', 'DELETED', 'UNDELIV', 'REJECTD']
                 # Do we need to forward the receipt to the original sender ?
                 if ((pdu_dlr_status in success_states and
                              registered_delivery_receipt == 'SMSC_DELIVERY_RECEIPT_REQUESTED') or
